Added node linking an hovering
Nodes can now be linked by cicking on the right circle of the node, and then be linked to another by clicking on the left circle of that node. The wire is a bezier curve and the curvature can easily be changed. Self-linking is disabled and when the mouse on the node, the wire doesn't get drawn. Also if you're making a link, and you select another node (not clicking on the left circle), the link gets broken. Drawing and re-calculating is limited by if statements and booleans. E.g. only if the node is selected, only if it wants to make a connection (drawing it with your mouse) .